C++ Limitations in Visual Studio:     Visual studio does not support defaulted move constructors.     http://msdn.microsoft.com/en-us/library/hh567368.aspx     http://msdn.microsoft.com/en-us/library/dn457344.aspx
    Visual Studio doesn't support noexcept

    Visual Studio doesn't support constexpr

I littered the source tree with #ifdef _WIN32, but I don't think this is a maintainable way to deal with this. It might be better to simply wait for Visual Studio to be more standards compliant.

To build open the Visual Studio 2013 Command prompt type:
c:/clane/win32/clane/> msbuild clane.sln